+++ ukat/mapping/t1.py
@@ -344,7 +344,7 @@ class T1:
                                                affine=self.affine)
                     nib.save(r1_nifti, base_path + '_r1_map.nii.gz')
                 elif result == 'mask':
-                    mask_nifti = nib.Nifti1Image(self.mask.astype(int),
+                    mask_nifti = nib.Nifti1Image(self.mask.astype(np.uint16),
                                                  affine=self.affine)
                     nib.save(mask_nifti, base_path + '_mask.nii.gz')
         else:
